Split Azure nodepool types into their own file
